Navigation bar shifted to HomeFeedActivity
