+++ app/src/main/java/com/example/android/codeforces/Activities/MainActivity.java
@@ -23,8 +23,6 @@ public class MainActivity extends AppCompatActivity {
 
     private TextInputEditText etPreferredHandle;
     private ImageView ivSubmit;
-    private Toolbar toolbar;
-    private DrawerLayout drawerLayout;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -33,16 +31,6 @@ public class MainActivity extends AppCompatActivity {
 
         etPreferredHandle = findViewById(R.id.etPreferredHandle);
         ivSubmit = findViewById(R.id.ivSubmit);
-        drawerLayout = findViewById(R.id.dashboardNavigationDrawer);
-        toolbar = findViewById(R.id.main_toolbar);
-        toolbar.setNavigationIcon(R.drawable.ic_nav_btn);
-
-        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                drawerLayout.openDrawer(GravityCompat.START);
-            }
-        });
 
         etPreferredHandle.setOnEditorActionListener(
                 new TextView.OnEditorActionListener() {
@@ -75,13 +63,5 @@ public class MainActivity extends AppCompatActivity {
                         }
                     }
                 });
-    }
-
-    @Override
-    public void onBackPressed() {
-        if (drawerLayout.isDrawerOpen(GravityCompat.START))
-            drawerLayout.closeDrawer(GravityCompat.START);
-        else
-            super.onBackPressed();
     }
 }
